feat cookiebanner as component
